Module tiff::bytecast

Trivial, internal byte transmutation.

A dependency like bytemuck would give us extra assurance of the safety but overall would not reduce the amount of total unsafety. We don’t use it in the interface where the traits would really become useful.

SAFETY: These are benign casts as we apply them to fixed size integer types only. All of them are naturally aligned, valid for all bit patterns and their alignment is surely at most their size (we assert the latter fact since it is ‘implementation defined’ if following the letter of the unsafe code guidelines).

TODO: Would like to use std-lib here.

Macros

integral_slice_as_bytes 🔒

Functions

f32_as_ne_bytes 🔒

f32_as_ne_mut_bytes 🔒

f64_as_ne_bytes 🔒

f64_as_ne_mut_bytes 🔒

i8_as_ne_bytes 🔒

i8_as_ne_mut_bytes 🔒

i16_as_ne_bytes 🔒

i16_as_ne_mut_bytes 🔒

i32_as_ne_bytes 🔒

i32_as_ne_mut_bytes 🔒

i64_as_ne_bytes 🔒

i64_as_ne_mut_bytes 🔒

u16_as_ne_bytes 🔒

u16_as_ne_mut_bytes 🔒

u32_as_ne_bytes 🔒

u32_as_ne_mut_bytes 🔒

u64_as_ne_bytes 🔒

u64_as_ne_mut_bytes 🔒