tiff::decoder

Struct Decoder

pub struct Decoder<R>
where
    R: Read + Seek,
{ /* private fields */ }

The representation of a TIFF decoder

Currently does not support decoding of interlaced images

Implementations

impl<R: Read + Seek> Decoder<R>

pub fn new(r: R) -> TiffResult<Decoder<R>>

pub fn with_limits(self, limits: Limits) -> Decoder<R>

pub fn dimensions(&mut self) -> TiffResult<(u32, u32)>

pub fn colortype(&mut self) -> TiffResult<ColorType>

pub fn ifd_pointer(&mut self) -> Option<IfdPointer>

The offset of the directory representing the current image.

pub fn seek_to_image(&mut self, ifd_index: usize) -> TiffResult<()>

Loads the IFD at the specified index in the list, if one exists

pub fn next_image(&mut self) -> TiffResult<()>

Reads in the next image. If there is no further image in the TIFF file a format error is returned. To determine whether there are more images call TIFFDecoder::more_images instead.

pub fn more_images(&self) -> bool

Returns true if there is at least one more image available.

pub fn byte_order(&self) -> ByteOrder

Returns the byte_order of the file.

pub fn read_ifd_offset(&mut self) -> Result<u64, Error>

pub fn inner(&mut self) -> &mut R

Returns a mutable reference to the stream being decoded.

pub fn read_byte(&mut self) -> Result<u8, Error>

Reads a TIFF byte value

pub fn read_short(&mut self) -> Result<u16, Error>

Reads a TIFF short value

pub fn read_sshort(&mut self) -> Result<i16, Error>

Reads a TIFF sshort value

pub fn read_long(&mut self) -> Result<u32, Error>

Reads a TIFF long value

pub fn read_slong(&mut self) -> Result<i32, Error>

Reads a TIFF slong value

pub fn read_float(&mut self) -> Result<f32, Error>

Reads a TIFF float value

pub fn read_double(&mut self) -> Result<f64, Error>

Reads a TIFF double value

pub fn read_long8(&mut self) -> Result<u64, Error>

pub fn read_slong8(&mut self) -> Result<i64, Error>

pub fn read_string(&mut self, length: usize) -> TiffResult<String>

Reads a string

pub fn read_offset(&mut self) -> TiffResult<[u8; 4]>

Reads a TIFF IFA offset/value field

pub fn read_offset_u64(&mut self) -> Result<[u8; 8], Error>

Reads a TIFF IFA offset/value field

pub fn goto_offset(&mut self, offset: u32) -> Result<()>

Moves the cursor to the specified offset

pub fn goto_offset_u64(&mut self, offset: u64) -> Result<()>

pub fn read_directory(&mut self, ptr: IfdPointer) -> TiffResult<Directory>

Read a tag-entry map from a known offset.

A TIFF Directory, aka. image file directory aka. IFD, refers to a map from tags–identified by a u16–to a typed vector of elements. It is encoded as a list of ascending tag values with the offset and type of their corresponding values. The semantic interpretations of a tag and its type requirements depend on the context of the directory. The main image directories, those iterated over by the Decoder after construction, are represented by Tag and ifd::Value. Other forms are EXIF and GPS data as well as thumbnail Sub-IFD representations associated with each image file.

This method allows the decoding of a directory from an arbitrary offset in the image file with no specific semantic interpretation. Such an offset is usually found as the value of a tag, e.g. Tag::SubIfd, Tag::ExifDirectory, Tag::GpsDirectory and recovered from the associated value by ifd::Value::into_ifd_pointer.

The library will not verify whether the offset overlaps any other directory or would form a cycle with any other directory when calling this method. This will modify the position of the reader, i.e. continuing with direct reads at a later point will require going back with Self::goto_offset.

pub fn get_chunk_type(&self) -> ChunkType

The chunk type (Strips / Tiles) of the image

pub fn strip_count(&mut self) -> TiffResult<u32>

Number of strips in image

pub fn tile_count(&mut self) -> TiffResult<u32>

Number of tiles in image

pub fn read_chunk_to_buffer( &mut self, buffer: DecodingBuffer<'_>, chunk_index: u32, output_width: usize, ) -> TiffResult<()>

pub fn image_chunk_buffer_layout( &mut self, chunk_index: u32, ) -> TiffResult<BufferLayoutPreference>

Returns the layout preferred to read the specified chunk with Self::read_chunk_bytes.

Returns the layout without being specific as to the underlying type for forward compatibility. Note that, in general, a TIFF may contain an almost arbitrary number of channels of individual bit length and format each.

See Self::colortype to describe the sample types.

Bugs

When the image is stored as a planar configuration, this method will currently only indicate the layout needed to read the first data plane. This will be fixed in a future major version of tiff.

pub fn read_chunk(&mut self, chunk_index: u32) -> TiffResult<DecodingResult>

Read the specified chunk (at index chunk_index) and return the binary data as a Vector.

Bugs

When the image is stored as a planar configuration, this method will currently only read the first sample’s plane. This will be fixed in a future major version of tiff.

pub fn read_chunk_bytes( &mut self, chunk_index: u32, buffer: &mut [u8], ) -> TiffResult<()>

Read the specified chunk (at index chunk_index) into an allocated buffer.

Returns a TiffError::UsageError if the chunk is smaller than the size indicated with a call to Self::image_chunk_buffer_layout. Note that the alignment may be arbitrary, but an alignment smaller than the preferred alignment may perform worse.

Bugs

When the image is stored as a planar configuration, this method will currently only read the first sample’s plane. This will be fixed in a future major version of tiff.

pub fn chunk_dimensions(&self) -> (u32, u32)

Returns the default chunk size for the current image. Any given chunk in the image is at most as large as the value returned here. For the size of the data (chunk minus padding), use chunk_data_dimensions.

pub fn chunk_data_dimensions(&self, chunk_index: u32) -> (u32, u32)

Returns the size of the data in the chunk with the specified index. This is the default size of the chunk, minus any padding.

pub fn image_buffer_layout(&mut self) -> TiffResult<BufferLayoutPreference>

Returns the preferred buffer required to read the whole image with Self::read_image_bytes.

Returns the layout without being specific as to the underlying type for forward compatibility. Note that, in general, a TIFF may contain an almost arbitrary number of channels of individual bit length and format each.

See Self::colortype to describe the sample types.

Bugs

When the image is stored as a planar configuration, this method will currently only indicate the layout needed to read the first data plane. This will be fixed in a future major version of tiff.

pub fn read_image(&mut self) -> TiffResult<DecodingResult>

Decodes the entire image and return it as a Vector

Bugs

When the image is stored as a planar configuration, this method will currently only read the first sample’s plane. This will be fixed in a future major version of tiff.

pub fn read_image_bytes(&mut self, buffer: &mut [u8]) -> TiffResult<()>

Decodes the entire image into a provided buffer.

Returns a TiffError::UsageError if the chunk is smaller than the size indicated with a call to Self::image_buffer_layout. Note that the alignment may be arbitrary, but an alignment smaller than the preferred alignment may perform worse.

Bugs

When the image is stored as a planar configuration, this method will currently only read the first sample’s plane. This will be fixed in a future major version of tiff.

pub fn read_directory_tags<'ifd>( &'ifd mut self, ifd: &'ifd Directory, ) -> IfdDecoder<'ifd>

Prepare reading values for tags of a given directory.

Examples

This method may be used to read the values of tags in directories that have been previously read with Decoder::read_directory.

use tiff::decoder::Decoder;
use tiff::tags::Tag;

let mut decoder = Decoder::new(&mut data).unwrap();
let sub_ifds = decoder.get_tag(Tag::SubIfd)?.into_ifd_vec()?;

for ifd in sub_ifds {
    let subdir = decoder.read_directory(ifd)?;
    let subfile = decoder.read_directory_tags(&subdir).find_tag(Tag::SubfileType)?;
    // omitted: handle the subfiles, e.g. thumbnails
}

pub fn find_tag(&mut self, tag: Tag) -> TiffResult<Option<Value>>

Tries to retrieve a tag from the current image directory. Return Ok(None) if the tag is not present.

pub fn find_tag_unsigned<T: TryFrom<u64>>( &mut self, tag: Tag, ) -> TiffResult<Option<T>>

Tries to retrieve a tag in the current image directory and convert it to the desired unsigned type.

pub fn find_tag_unsigned_vec<T: TryFrom<u64>>( &mut self, tag: Tag, ) -> TiffResult<Option<Vec<T>>>

Tries to retrieve a vector of all a tag’s values and convert them to the desired unsigned type.

pub fn get_tag_unsigned<T: TryFrom<u64>>(&mut self, tag: Tag) -> TiffResult<T>

Tries to retrieve a tag from the current image directory and convert it to the desired unsigned type. Returns an error if the tag is not present.

pub fn get_tag(&mut self, tag: Tag) -> TiffResult<Value>

Tries to retrieve a tag from the current image directory. Returns an error if the tag is not present

pub fn get_tag_u32(&mut self, tag: Tag) -> TiffResult<u32>

pub fn get_tag_u64(&mut self, tag: Tag) -> TiffResult<u64>

pub fn get_tag_f32(&mut self, tag: Tag) -> TiffResult<f32>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_f64(&mut self, tag: Tag) -> TiffResult<f64>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u32_vec(&mut self, tag: Tag) -> TiffResult<Vec<u32>>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u16_vec(&mut self, tag: Tag) -> TiffResult<Vec<u16>>

pub fn get_tag_u64_vec(&mut self, tag: Tag) -> TiffResult<Vec<u64>>

pub fn get_tag_f32_vec(&mut self, tag: Tag) -> TiffResult<Vec<f32>>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_f64_vec(&mut self, tag: Tag) -> TiffResult<Vec<f64>>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u8_vec(&mut self, tag: Tag) -> TiffResult<Vec<u8>>

Tries to retrieve a tag and convert it to a 8bit vector.

pub fn get_tag_ascii_string(&mut self, tag: Tag) -> TiffResult<String>

Tries to retrieve a tag and convert it to a ascii vector.

pub fn tag_iter( &mut self, ) -> impl Iterator<Item = TiffResult<(Tag, Value)>> + '_