tiff

Module encoder

Modules

colortype

compression

Structs

DirectoryEncoder

Low level interface to encode ifd directories.

DirectoryOffset

The offset of an encoded directory in the file.

Ifd

Type to represent tiff values of type IFD

Ifd8

Type to represent tiff values of type IFD8

ImageEncoder

Type to encode images strip by strip.

Rational

Type to represent tiff values of type RATIONAL

SRational

Type to represent tiff values of type SRATIONAL

TiffEncoder

Encoder for Tiff and BigTiff files.

TiffKindBig

Create a BigTiff file.

TiffKindStandard

Create a standard Tiff file.

Enums

Compression

Traits

TiffKind

Trait to abstract over Tiff/BigTiff differences.

TiffValue

Trait for types that can be encoded in a tiff file

Type Aliases

DeflateLevel

Predictor

Type of prediction to prepare the image with.