Enum tiff::TiffFormatError

#[non_exhaustive]
pub enum TiffFormatError {

    TiffSignatureNotFound,
    TiffSignatureInvalid,
    ImageFileDirectoryNotFound,
    InconsistentSizesEncountered,
    UnexpectedCompressedData {
        actual_bytes: usize,
        required_bytes: usize,
    },
    InconsistentStripSamples {
        actual_samples: usize,
        required_samples: usize,
    },
    InvalidDimensions(u32, u32),
    InvalidTag,
    InvalidTagValueType(Tag),
    RequiredTagNotFound(Tag),
    UnknownPredictor(u16),
    UnknownPlanarConfiguration(u16),
    ByteExpected(Value),
    UnsignedIntegerExpected(Value),
    SignedIntegerExpected(Value),
    Format(String),
    RequiredTagEmpty(Tag),
    StripTileTagConflict,
    CycleInOffsets,
    JpegDecoder(JpegDecoderError),
    SamplesPerPixelIsZero,

}

The image is not formatted properly.

This indicates that the encoder producing the image might behave incorrectly or that the input file has been corrupted.

The list of variants may grow to incorporate errors of future features. Matching against this exhaustively is not covered by interface stability guarantees.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

TiffSignatureNotFound

TiffSignatureInvalid

ImageFileDirectoryNotFound

InconsistentSizesEncountered

UnexpectedCompressedData

Fields

actual_bytes: usize

required_bytes: usize

InconsistentStripSamples

Fields

actual_samples: usize

required_samples: usize

InvalidDimensions(u32, u32)

InvalidTag

InvalidTagValueType(Tag)

RequiredTagNotFound(Tag)

UnknownPredictor(u16)

UnknownPlanarConfiguration(u16)

ByteExpected(Value)

UnsignedIntegerExpected(Value)

SignedIntegerExpected(Value)

Format(String)

RequiredTagEmpty(Tag)

StripTileTagConflict

CycleInOffsets

JpegDecoder(JpegDecoderError)

SamplesPerPixelIsZero