tiff

Enum UsageError

#[non_exhaustive]
pub enum UsageError {

    InvalidChunkType(ChunkType, ChunkType),
    InvalidChunkIndex(u32),
    InvalidPlaneIndex(u16),
    InvalidCodingUnit(u32, u32),
    PredictorCompressionMismatch,
    PredictorIncompatible,
    PredictorUnavailable,
    InsufficientOutputBufferSize {
        needed: usize,
        provided: usize,
    },
    InsufficientOutputRowStride {
        needed: usize,
        requested: usize,
    },
    ZeroIfdPointer,
    ReconfiguredAfterImageWrite,
    ByteOrderMismatch,
    MismatchedEntryLength {
        ty: Type,
        found: usize,
    },

}

User attempted to use the Decoder in a way that is incompatible with a specific image.

For example: attempting to read a tile from a stripped image.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InvalidChunkType(ChunkType, ChunkType)

InvalidChunkIndex(u32)

InvalidPlaneIndex(u16)

InvalidCodingUnit(u32, u32)

PredictorCompressionMismatch

PredictorIncompatible

PredictorUnavailable

InsufficientOutputBufferSize

Fields

needed: usize

provided: usize

InsufficientOutputRowStride

Fields

needed: usize

requested: usize

ZeroIfdPointer

ReconfiguredAfterImageWrite

ByteOrderMismatch

MismatchedEntryLength

Fields

ty: Type

found: usize