Struct tiff::EndianFile

pub struct EndianFile { /* fields omitted */ }

A helper structure that provides convenience methods to write to a fs::File, being aware of the file's Endianness.

Methods

impl EndianFile

pub fn write_u8(&mut self, n: u8) -> Result<()>

Writes a u8 to the file.

pub fn write_u16(&mut self, n: u16) -> Result<()>

Writes a u16 to the file.

pub fn write_u32(&mut self, n: u32) -> Result<()>

Writes a u32 to the file.

pub fn write_i8(&mut self, n: i8) -> Result<()>

Writes a i8 to the file.

pub fn write_i16(&mut self, n: i16) -> Result<()>

Writes a i16 to the file.

pub fn write_i32(&mut self, n: i32) -> Result<()>

Writes a i32 to the file.

pub fn write_f32(&mut self, n: f32) -> Result<()>

Writes a f32 to the file.

pub fn write_f64(&mut self, n: f64) -> Result<()>

Writes a f64 to the file.

pub fn write_arbitrary_byte(&mut self) -> Result<()>

Writes an arbitraty byte to the file.

This is useful when there is need to write an extra byte to guarantee that all offsets are even but that byte doesn't hold any information.