Module tiff::tiff_type

Representations of TIFF field data types.

Each type comes with convenience functions in order to facilitate its use.

Every TIFF data type implements TiffType in order to be usable in the crate.

Structs

ASCII
8-bit byte that contains a 7-bit ASCII code.

BYTE
8-bit unsigned integer.

DOUBLE
Double precision (8-byte) IEEE format.

FLOAT
Single precision (4-byte) IEEE format.

LONG
32-bit (4-byte) unsigned integer.

RATIONAL
Two LONGs representing, respectively, the numerator and the denominator of a fraction.

SBYTE
8-bit signed (twos-complement) integer.

SHORT
16-bit (2-byte) unsigned integer.

SLONG
32-bit (4-byte) signed (twos-complement) integer.

SRATIONAL
Two SLONGs representing, respectively, the numerator and the denominator of a fraction.

SSHORT
16-bit (2-byte) signed (twos-complement) integer.

UNDEFINED
8-bit byte that may contain anything, depending on the definition of the field.

Traits

TiffType
A TIFF field type of data.