Trait tiff::tiff_type::TiffType

pub trait TiffType {
    fn id() -> u16;

    fn size() -> u32;

    fn write_to(self, file: &mut EndianFile) -> Result<()>;
}

A TIFF field type of data.

If a new type is created (or is present in an extension of TIFF) that this crate doesn't support yet, it can be easily implemented by implementing this trait.

Required Methods

fn id() -> u16

The TIFF 16-bit code that identifies the type.

fn size() -> u32

The number of bytes occupied by a single value of this type.

fn write_to(self, file: &mut EndianFile) -> Result<()>

The function that writes this type to a given EndianFile.

Panics

Will panic if the number of bytes written to the file is different than the number of bytes specified in size().