Trait tiff::Datablock

pub trait Datablock {
    type Allocated: AllocatedDatablock;
    fn allocate(self, c: &mut Cursor) -> Self::Allocated;
}

Represents a block of data in the file pointed to by a field value, but that isn't part of the field itself.

This includes, for example, subIFDs and image strips.

In most cases, it is simpler to implement SimpleDatablock instead, which will automatically implement both Datablock and its allocated form, AllocatedDatablock. Implementing Datablock directly is only useful if there is need to do any special operation during the allocation (for example, to track the position of the block in the file in order to write offsets).

Associated Types

type Allocated: AllocatedDatablock

The allocated form of this Datablock.

It is valid for a structure to be both its normal and allocated form, as long as it doens't need to store any different information after allocation. In that case, type Allocated = Self.

However, if you find yourself doing that, consider using SimpleDataBlock instead .

Required Methods

fn allocate(self, c: &mut Cursor) -> Self::Allocated

Allocates this Datablock, moving the Cursor forwards exactly the same number of bytes as those that will be written in write_to(self, &mut EndianFile). Returns its allocated form, Self::Allocated.

Panics

Failing to allocate the exact same number of bytes that will be written to the file will panic.