Crate tiff

Decoding and Encoding of TIFF Images

TIFF (Tagged Image File Format) is a versatile image format that supports lossless and lossy compression.

Related Links

https://web.archive.org/web/20210108073850/https://www.adobe.io/open/standards/TIFF.html - The TIFF specification

Modules

decoder

encoder

tags

Enums

ColorType

An enumeration over supported color types and their bit depths

TiffError

Tiff error kinds.

TiffFormatError

The image is not formatted properly.

TiffUnsupportedError

The Decoder does not support features required by the image.

UsageError

User attempted to use the Decoder in a way that is incompatible with a specific image.

Type Definitions

TiffResult

Result of an image decoding/encoding process