Crate tiff

Decoding and Encoding of TIFF Images

TIFF (Tagged Image File Format) is a versatile image format that supports lossless and lossy compression.

Related Links

http://partners.adobe.com/public/developer/tiff/index.html - The TIFF specification

Modules

decoder

encoder

error

tags

Structs

InflateError

Decompression failed due to faulty compressed data.

Enums

ColorType

An enumeration over supported color types and their bit depths

TiffError

Tiff error kinds.

TiffFormatError

The image is not formatted properly.

TiffUnsupportedError

The Decoder does not support features required by the image.

Type Definitions

TiffResult

Result of an image decoding/encoding process