tiff::decoder
Struct Decoder

pub struct Decoder<R>
where
    R: Read + Seek,
{ /* private fields */ }

The representation of a TIFF decoder

Currently does not support decoding of interlaced images

Implementations

impl<R: Read + Seek> Decoder<R>

pub fn new(r: R) -> TiffResult<Decoder<R>>

Create a new decoder that decodes from the stream r

pub fn with_limits(self, limits: Limits) -> Decoder<R>

pub fn dimensions(&mut self) -> TiffResult<(u32, u32)>

pub fn colortype(&mut self) -> TiffResult<ColorType>

pub fn seek_to_image(&mut self, ifd_index: usize) -> TiffResult<()>

Loads the IFD at the specified index in the list, if one exists

pub fn init(self) -> TiffResult<Decoder<R>>

👎Deprecated: Never should have been public. Only use Decoder::new()

Reset the decoder.

pub fn next_image(&mut self) -> TiffResult<()>

Reads in the next image. If there is no further image in the TIFF file a format error is returned. To determine whether there are more images call TIFFDecoder::more_images instead.

pub fn more_images(&self) -> bool

Returns true if there is at least one more image available.

pub fn byte_order(&self) -> ByteOrder

Returns the byte_order

pub fn read_ifd_offset(&mut self) -> Result<u64, Error>

pub fn read_byte(&mut self) -> Result<u8, Error>

Reads a TIFF byte value

pub fn read_short(&mut self) -> Result<u16, Error>

Reads a TIFF short value

pub fn read_sshort(&mut self) -> Result<i16, Error>

Reads a TIFF sshort value

pub fn read_long(&mut self) -> Result<u32, Error>

Reads a TIFF long value

pub fn read_slong(&mut self) -> Result<i32, Error>

Reads a TIFF slong value

pub fn read_float(&mut self) -> Result<f32, Error>

Reads a TIFF float value

pub fn read_double(&mut self) -> Result<f64, Error>

Reads a TIFF double value

pub fn read_long8(&mut self) -> Result<u64, Error>

pub fn read_slong8(&mut self) -> Result<i64, Error>

pub fn read_string(&mut self, length: usize) -> TiffResult<String>

Reads a string

pub fn read_offset(&mut self) -> TiffResult<[u8; 4]>

Reads a TIFF IFA offset/value field

pub fn read_offset_u64(&mut self) -> Result<[u8; 8], Error>

Reads a TIFF IFA offset/value field

pub fn goto_offset(&mut self, offset: u32) -> Result<()>

Moves the cursor to the specified offset

pub fn goto_offset_u64(&mut self, offset: u64) -> Result<()>

pub fn find_tag(&mut self, tag: Tag) -> TiffResult<Option<Value>>

Tries to retrieve a tag. Return Ok(None) if the tag is not present.

pub fn find_tag_unsigned<T: TryFrom<u64>>( &mut self, tag: Tag, ) -> TiffResult<Option<T>>

Tries to retrieve a tag and convert it to the desired unsigned type.

pub fn find_tag_unsigned_vec<T: TryFrom<u64>>( &mut self, tag: Tag, ) -> TiffResult<Option<Vec<T>>>

Tries to retrieve a vector of all a tag’s values and convert them to the desired unsigned type.

pub fn get_tag_unsigned<T: TryFrom<u64>>(&mut self, tag: Tag) -> TiffResult<T>

Tries to retrieve a tag and convert it to the desired unsigned type. Returns an error if the tag is not present.

pub fn get_tag(&mut self, tag: Tag) -> TiffResult<Value>

Tries to retrieve a tag. Returns an error if the tag is not present

pub fn get_tag_u32(&mut self, tag: Tag) -> TiffResult<u32>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u64(&mut self, tag: Tag) -> TiffResult<u64>

pub fn get_tag_f32(&mut self, tag: Tag) -> TiffResult<f32>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_f64(&mut self, tag: Tag) -> TiffResult<f64>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u32_vec(&mut self, tag: Tag) -> TiffResult<Vec<u32>>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u16_vec(&mut self, tag: Tag) -> TiffResult<Vec<u16>>

pub fn get_tag_u64_vec(&mut self, tag: Tag) -> TiffResult<Vec<u64>>

pub fn get_tag_f32_vec(&mut self, tag: Tag) -> TiffResult<Vec<f32>>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_f64_vec(&mut self, tag: Tag) -> TiffResult<Vec<f64>>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u8_vec(&mut self, tag: Tag) -> TiffResult<Vec<u8>>

Tries to retrieve a tag and convert it to a 8bit vector.

pub fn get_tag_ascii_string(&mut self, tag: Tag) -> TiffResult<String>

Tries to retrieve a tag and convert it to a ascii vector.

pub fn get_chunk_type(&self) -> ChunkType

The chunk type (Strips / Tiles) of the image

pub fn strip_count(&mut self) -> TiffResult<u32>

Number of strips in image

pub fn tile_count(&mut self) -> TiffResult<u32>

Number of tiles in image

pub fn read_jpeg(&mut self) -> TiffResult<DecodingResult>

👎Deprecated: Use read_image instead

pub fn read_strip_to_buffer( &mut self, buffer: DecodingBuffer<'_>, ) -> TiffResult<()>

👎Deprecated: Use read_chunk_to_buffer instead

pub fn read_chunk_to_buffer( &mut self, buffer: DecodingBuffer<'_>, chunk_index: u32, output_width: usize, ) -> TiffResult<()>

pub fn read_strip(&mut self) -> TiffResult<DecodingResult>

👎Deprecated: Use read_chunk instead

Read a single strip from the image and return it as a Vector. This method does not return information on the size of the strip read and so the caller must take care of the final strip in the image as the data read may be less than the expected strip size. This interface may change at a future date to reflect this.

pub fn read_tile(&mut self) -> TiffResult<DecodingResult>

👎Deprecated: Use read_chunk instead

Read a single tile from the image and return it as a Vector. This method does not return information on the size of the tile read and so the caller must take care of tiles on the rightmost column and the bottommost row in the image as the data read may be less than the expected tile size. This interface may change at a future date to reflect this.

pub fn read_chunk(&mut self, chunk_index: u32) -> TiffResult<DecodingResult>

Read the specified chunk (at index chunk_index) and return the binary data as a Vector.

pub fn chunk_dimensions(&self) -> (u32, u32)

Returns the default chunk size for the current image. Any given chunk in the image is at most as large as the value returned here. For the size of the data (chunk minus padding), use chunk_data_dimensions.

pub fn chunk_data_dimensions(&self, chunk_index: u32) -> (u32, u32)

Returns the size of the data in the chunk with the specified index. This is the default size of the chunk, minus any padding.

pub fn read_image(&mut self) -> TiffResult<DecodingResult>

Decodes the entire image and return it as a Vector