Struct tiff::decoder::Decoder

pub struct Decoder<R> where
    R: Read + Seek,  { /* fields omitted */ }

The representation of a TIFF decoder

Currently does not support decoding of interlaced images

Methods

impl<R: Read + Seek> Decoder<R>

pub fn new(r: R) -> TiffResult<Decoder<R>>

Create a new decoder that decodes from the stream r

pub fn dimensions(&mut self) -> TiffResult<(u32, u32)>

pub fn colortype(&mut self) -> TiffResult<ColorType>

pub fn init(self) -> TiffResult<Decoder<R>>

Initializes the decoder.

pub fn next_image(self) -> TiffResult<Decoder<R>>

Reads in the next image. If there is no further image in the TIFF file a format error is returned. To determine whether there are more images call TIFFDecoder::more_images instead.

pub fn more_images(&self) -> bool

Returns true if there is at least one more image available.

pub fn byte_order(&self) -> ByteOrder

Returns the byte_order

pub fn read_short(&mut self) -> Result<u16, Error>

Reads a TIFF short value

pub fn read_long(&mut self) -> Result<u32, Error>

Reads a TIFF long value

pub fn read_string(&mut self, length: usize) -> Result<String, FromUtf8Error>

Reads a string

pub fn read_offset(&mut self) -> Result<[u8; 4], Error>

Reads a TIFF IFA offset/value field

pub fn goto_offset(&mut self, offset: u32) -> Result<()>

Moves the cursor to the specified offset

pub fn find_tag(&mut self, tag: Tag) -> TiffResult<Option<Value>>

Tries to retrieve a tag. Return Ok(None) if the tag is not present.

pub fn find_tag_u32(&mut self, tag: Tag) -> TiffResult<Option<u32>>

Tries to retrieve a tag and convert it to the desired type.

pub fn find_tag_u32_vec(&mut self, tag: Tag) -> TiffResult<Option<Vec<u32>>>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag(&mut self, tag: Tag) -> TiffResult<Value>

Tries to retrieve a tag. Returns an error if the tag is not present

pub fn get_tag_u32(&mut self, tag: Tag) -> TiffResult<u32>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u32_vec(&mut self, tag: Tag) -> TiffResult<Vec<u32>>

Tries to retrieve a tag and convert it to the desired type.

pub fn read_image(&mut self) -> TiffResult<DecodingResult>

Decodes the entire image and return it as a Vector