Struct tiff::decoder::Decoder

pub struct Decoder<R> where
    R: Read + Seek,  {
    reader: SmartReader<R>,
    byte_order: ByteOrder,
    limits: Limits,
    next_ifd: Option<u32>,
    ifd: Option<Directory>,
    width: u32,
    height: u32,
    bits_per_sample: Vec<u8>,
    samples: u8,
    photometric_interpretation: PhotometricInterpretation,
    compression_method: CompressionMethod,
    strip_decoder: Option<StripDecodeState>,
}

The representation of a TIFF decoder

Currently does not support decoding of interlaced images

Fields
reader: SmartReader<R>byte_order: ByteOrderlimits: Limitsnext_ifd: Option<u32>ifd: Option<Directory>width: u32height: u32bits_per_sample: Vec<u8>samples: u8photometric_interpretation: PhotometricInterpretationcompression_method: CompressionMethodstrip_decoder: Option<StripDecodeState>
Implementations

impl<R: Read + Seek> Decoder<R>

pub fn new(r: R) -> TiffResult<Decoder<R>>

Create a new decoder that decodes from the stream r

pub fn with_limits(self, limits: Limits) -> Decoder<R>

pub fn dimensions(&mut self) -> TiffResult<(u32, u32)>

pub fn colortype(&mut self) -> TiffResult<ColorType>

fn read_header(&mut self) -> TiffResult<()>

pub fn init(self) -> TiffResult<Decoder<R>>

Initializes the decoder.

pub fn next_image(&mut self) -> TiffResult<()>

Reads in the next image. If there is no further image in the TIFF file a format error is returned. To determine whether there are more images call TIFFDecoder::more_images instead.

pub fn more_images(&self) -> bool

Returns true if there is at least one more image available.

pub fn byte_order(&self) -> ByteOrder

Returns the byte_order

pub fn read_byte(&mut self) -> Result<u8, Error>

Reads a TIFF byte value

pub fn read_short(&mut self) -> Result<u16, Error>

Reads a TIFF short value

pub fn read_sshort(&mut self) -> Result<i16, Error>

Reads a TIFF sshort value

pub fn read_long(&mut self) -> Result<u32, Error>

Reads a TIFF long value

pub fn read_slong(&mut self) -> Result<i32, Error>

Reads a TIFF slong value

pub fn read_string(&mut self, length: usize) -> TiffResult<String>

Reads a string

pub fn read_offset(&mut self) -> Result<[u8; 4], Error>

Reads a TIFF IFA offset/value field

pub fn goto_offset(&mut self, offset: u32) -> Result<()>

Moves the cursor to the specified offset

fn read_entry(&mut self) -> TiffResult<Option<(Tag, Entry)>>

Reads a IFD entry.

fn read_ifd(&mut self) -> TiffResult<Directory>

Reads the next IFD

pub fn find_tag(&mut self, tag: Tag) -> TiffResult<Option<Value>>

Tries to retrieve a tag. Return Ok(None) if the tag is not present.

pub fn find_tag_unsigned<T: TryFrom<u32>>(
&mut self,
tag: Tag
) -> TiffResult<Option<T>>

Tries to retrieve a tag and convert it to the desired unsigned type.

pub fn find_tag_unsigned_vec<T: TryFrom<u32>>(
&mut self,
tag: Tag
) -> TiffResult<Option<Vec<T>>>

Tries to retrieve a vector of all a tag’s values and convert them to the desired unsigned type.

pub fn get_tag_unsigned<T: TryFrom<u32>>(&mut self, tag: Tag) -> TiffResult<T>

Tries to retrieve a tag and convert it to the desired unsigned type. Returns an error if the tag is not present.

pub fn get_tag(&mut self, tag: Tag) -> TiffResult<Value>

Tries to retrieve a tag. Returns an error if the tag is not present

pub fn get_tag_u32(&mut self, tag: Tag) -> TiffResult<u32>

Tries to retrieve a tag and convert it to the desired type.

pub fn get_tag_u32_vec(&mut self, tag: Tag) -> TiffResult<Vec<u32>>

Tries to retrieve a tag and convert it to the desired type.

fn expand_strip<'a>(
&mut self,
buffer: DecodingBuffer<'a>,
offset: u32,
length: u32,
max_uncompressed_length: usize
) -> TiffResult<usize>

Decompresses the strip into the supplied buffer. Returns the number of bytes read.

pub fn strip_count(&mut self) -> TiffResult<u32>

Number of strips in image

fn initialize_strip_decoder(&mut self) -> TiffResult<()>

pub fn read_strip_to_buffer(
&mut self,
buffer: DecodingBuffer<'_>
) -> TiffResult<()>

fn result_buffer(&self, height: usize) -> TiffResult<DecodingResult>

pub fn read_strip(&mut self) -> TiffResult<DecodingResult>

Read a single strip from the image and return it as a Vector

pub fn read_image(&mut self) -> TiffResult<DecodingResult>

Decodes the entire image and return it as a Vector