tiff::encoder

Module compression

Structs

Deflate

The Deflate algorithm used to compress image data in TIFF files.

Lzw

The LZW algorithm used to compress image data in TIFF files.

Packbits

Compressor that uses the Packbits algorithm to compress bytes.

Uncompressed

The default algorithm which does not compress at all.

Enums

Compressor

An enum to store each compression algorithm.

DeflateLevel

The level of compression used by the Deflate algorithm. It allows trading compression ratio for compression speed.

Traits

Compression

An algorithm used for compression with associated enums and optional configurations.

CompressionAlgorithm

An algorithm used for compression