Trait tiff::encoder::compression::Compression

pub trait Compression: CompressionAlgorithm {
    const COMPRESSION_METHOD: CompressionMethod;

    // Required method
    fn get_algorithm(&self) -> Compressor;
}

An algorithm used for compression with associated enums and optional configurations.

Required Associated Constants

const COMPRESSION_METHOD: CompressionMethod

The corresponding tag to the algorithm.

Required Methods

fn get_algorithm(&self) -> Compressor

Method to optain a type that can store each variant of comression algorithm.

Object Safety

This trait is not object safe.