Trait tiff::encoder::TiffValue

pub trait TiffValue {
    const BYTE_LEN: u8;
    const FIELD_TYPE: Type;

    fn count(&self) -> usize;
    fn data(&self) -> Cow<'_, [u8]>;

    fn bytes(&self) -> usize { ... }
    fn write<W: Write>(&self, writer: &mut TiffWriter<W>) -> TiffResult<()> { ... }
}

Trait for types that can be encoded in a tiff file

Required Associated Constants

const BYTE_LEN: u8

const FIELD_TYPE: Type

Required Methods

fn count(&self) -> usize

fn data(&self) -> Cow<'_, [u8]>

Access this value as an contiguous sequence of bytes. If their is no trivial representation, allocate it on the heap.

Provided Methods

fn bytes(&self) -> usize

fn write<W: Write>(&self, writer: &mut TiffWriter<W>) -> TiffResult<()>

Write this value to a TiffWriter. While the default implementation will work in all cases, it may require unnecessary allocations. The written bytes of any custom implementation MUST be the same as yielded by self.data().