Enum tiff::TiffError

pub enum TiffError {
    FormatError(TiffFormatError),
    UnsupportedError(TiffUnsupportedError),
    IoError(Error),
    LimitsExceeded,
    IntSizeError,
    UsageError(UsageError),
}

Tiff error kinds.

Variants

FormatError(TiffFormatError)

The Image is not formatted properly.

UnsupportedError(TiffUnsupportedError)

The Decoder does not support features required by the image.

IoError(Error)

An I/O Error occurred while decoding the image.

LimitsExceeded

The Limits of the Decoder is exceeded.

IntSizeError

An integer conversion to or from a platform size failed, either due to limits of the platform size or limits of the format.

UsageError(UsageError)

The image does not support the requested operation